# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediChain API
  description: |
    MediChain REST API provides secure access to patient identity, medical records, 
    and access control functionality for a blockchain-based national health ID system.
    
    © 2025 Lukau Invasion (Pty) Ltd. All rights reserved.
    
    ## Authentication
    
    MediChain uses **wallet-based blockchain authentication** with SS58 addresses. 
    For API testing, include the `X-User-Id` header with a valid wallet address.
    
    ## Role-Based Access Control
    
    All endpoints implement RBAC with the following roles:
    - **Admin** - Full system access, user management
    - **Doctor** - Patient registration, medical record access/edit
    - **Nurse** - Patient registration, medical record access/edit
    - **LabTechnician** - Lab results upload, limited record access
    - **Pharmacist** - Prescription verification, medication records
    - **Patient** - Read-only access to own records
    
    ## Pagination
    
    List endpoints support pagination via query parameters:
    - `page` - Page number (1-indexed, default: 1)
    - `limit` - Items per page (default: 20, max: 100)
    
  version: "1.0.0"
  contact:
    name: MediChain Support
    email: support@medichain.ng
  license:
    name: Proprietary
    url: https://medichain.ng/license
servers:
  - url: http://localhost:8090
    description: Local development server
  - url: https://api.medichain.ng
    description: Production server
...
security:
  - WalletAuth: []
paths:
...
  /api/nfc/generate:
    post:
      tags: [NFC Cards]
      summary: Generate NFC card
      description: Generate a new NFC health card for a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_id]
              properties:
                patient_id:
                  type: string
      responses:
        '201':
          description: NFC card generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateNFCCardResponse'
...
  /api/nfc/cards:
    get:
      tags: [NFC Cards]
      summary: List NFC cards
      description: Get paginated list of all NFC cards (Admin only)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: NFC cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNFCCards'
        '403':
          description: Only Admin can list cards
...
components:
  securitySchemes:
    WalletAuth:
      type: apiKey
      in: header
      name: X-User-Id
      description: SS58 encoded wallet address
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number (1-indexed)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    GenerateNFCCardResponse:
      type: object
      properties:
        success:
          type: boolean
        card_id:
          type: string
        card_hash:
          type: string
        qr_code_base64:
          type: string
    CardInfo:
      type: object
      properties:
        card_id:
          type: string
        patient_id:
          type: string
        card_hash:
          type: string
        national_id_type:
          type: string
        status:
          type: string
        created_at:
          type: integer
        last_used_at:
          type: integer
    PaginatedNFCCards:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardInfo'
        total:
          type: integer
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...